+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Correct setting name on AzureAD Tenant backend
 - Introduce access token expired threshold of 5 seconds by default
 - Delete partial token from session if still present
+- Use `userPrincipalName` to set `username` and `email` accordingly
 
 ## [1.7.0](https://github.com/python-social-auth/social-core/releases/tag/1.7.0) - 2018-02-20